refactor: xmodule/error_module.py -> xmodule/error_block.py
