+++ lms/djangoapps/ccx/models.py
@@ -14,7 +14,7 @@ from lazy import lazy
 from opaque_keys.edx.django.models import CourseKeyField, UsageKeyField
 from pytz import utc
 
-from xmodule.error_module import ErrorBlock
+from xmodule.error_block import ErrorBlock
 from xmodule.modulestore.django import modulestore
 
 log = logging.getLogger("edx.ccx")
