+++ lms/djangoapps/courseware/rules.py
@@ -20,7 +20,7 @@ from openedx.core.djangoapps.enrollments.api import is_enrollment_valid_for_proc
 from common.djangoapps.student.models import CourseAccessRole
 from common.djangoapps.student.roles import CourseRole, OrgRole
 from xmodule.course_block import CourseBlock  # lint-amnesty, pylint: disable=wrong-import-order
-from xmodule.error_module import ErrorBlock  # lint-amnesty, pylint: disable=wrong-import-order
+from xmodule.error_block import ErrorBlock  # lint-amnesty, pylint: disable=wrong-import-order
 
 
 from .access import has_access
