+++ setup.py
@@ -16,7 +16,7 @@ XBLOCKS = [
     "custom_tag_template = xmodule.template_module:CustomTagTemplateBlock",
     "discuss = xmodule.template_module:TranslateCustomTagBlock",
     "discussion = xmodule.discussion_block:DiscussionXBlock",
-    "error = xmodule.error_module:ErrorBlock",
+    "error = xmodule.error_block:ErrorBlock",
     "hidden = xmodule.hidden_module:HiddenDescriptor",
     "html = xmodule.html_module:HtmlBlock",
     "image = xmodule.template_module:TranslateCustomTagBlock",
