+++ xmodule/modulestore/split_mongo/caching_descriptor_system.py
@@ -9,7 +9,7 @@ from opaque_keys.edx.locator import BlockUsageLocator, DefinitionLocator, LocalI
 from xblock.fields import ScopeIds
 from xblock.runtime import KeyValueStore, KvsFieldData
 
-from xmodule.error_module import ErrorBlock
+from xmodule.error_block import ErrorBlock
 from xmodule.errortracker import exc_info_to_str
 from xmodule.library_tools import LibraryToolsService
 from xmodule.mako_module import MakoDescriptorSystem
